fix: moving from one app to another, tests outstanding
